Changed Vector3d to Vector3i in virtual function

+++ include/bundleadjust/Dataloader.h
@@ -15,7 +15,7 @@ public:
     inline virtual int getNumPoints() const = 0;
     inline virtual std::vector<cv::Point2f> getObservations() const = 0;
     inline virtual int getNumObservations() const = 0;
-    inline virtual Eigen::Vector3d getPointColor(int index) const = 0;
+    inline virtual Eigen::Vector3i getPointColor(int index) const = 0;
     inline virtual int getNumFrames() const = 0;
     inline virtual bool isColorAvailable() const = 0;
     inline virtual bool isDepthAvailable() const = 0;
